[refactor] round domain으로 이동
game manager 코드 리팩토링
